+++ app/layout.tsx
@@ -12,7 +12,9 @@ export const metadata: Metadata = {
   metadataBase: new URL(
     basePath
       ? "https://erict16.github.io/oltc-selector"
-      : "http://127.0.0.1:3000",
+      : process.env.VERCEL_PROJECT_PRODUCTION_URL
+        ? `https://${process.env.VERCEL_PROJECT_PRODUCTION_URL}`
+        : "http://127.0.0.1:3000",
   ),
   icons: {
     icon: [{ url: `${basePath}/favicon.svg`, type: "image/svg+xml" }],